Slim code, delete SnoopPrecompile
