+++ Project.toml
@@ -19,10 +19,14 @@ YAML = "ddb6d928-2868-570f-bddf-ab3f9cf99eb6"
 [weakdeps]
 Cthulhu = "f68482b8-f384-11e8-15f7-abe071a5a75f"
 JET = "c3a54625-cd67-489e-a8e7-0a5a0ff4e31b"
+PrettyTables = "08abe8d2-0d0c-5749-adfa-8a2ac140af0d"
+PyPlot = "d330b81b-6aea-500a-939a-2ce795aea3ee"
 
 [extensions]
 CthulhuExt = "Cthulhu"
 JETExt = ["JET", "Cthulhu"]
+SCPrettyTablesExt = "PrettyTables"
+SCPyPlotExt = "PyPlot"
 
 [compat]
 AbstractTrees = "0.4"
