--- SnoopPrecompile/Project.toml
@@ -1,18 +1,0 @@
-name = "SnoopPrecompile"
-uuid = "66db9d55-30c0-4569-8b51-7e840670fc0c"
-authors = ["Tim Holy <tim.holy@gmail.com> and contributors"]
-version = "1.0.3"
-
-[deps]
-Preferences = "21216c6a-2e73-6563-6e65-726566657250"
-
-[compat]
-Preferences = "1"
-julia = "1"
-
-[extras]
-Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
-UUIDs = "cf7118a7-6976-5b1a-9a39-7adc72f591a4"
-
-[targets]
-test = ["Test", "UUIDs"]
