--- SnoopPrecompile/README.md
@@ -1,5 +1,0 @@
-# SnoopPrecompile
-
-SnoopPrecompile is a small dependency used to effectively precompile code needed by your package, particularly on Julia 1.8 and higher.
-
-See the [documentation](https://timholy.github.io/SnoopCompile.jl/dev/snoop_pc/) for details.
