--- SnoopPrecompile/test/SnoopPC_B/Project.toml
@@ -1,7 +1,0 @@
-name = "SnoopPC_B"
-uuid = "d38b61e7-59a2-4ef9-b4d3-320bdc69b817"
-authors = ["Tim Holy <tim.holy@gmail.com>"]
-version = "0.1.0"
-
-[deps]
-SnoopPrecompile = "66db9d55-30c0-4569-8b51-7e840670fc0c"
