--- SnoopPrecompile/test/SnoopPC_B/src/SnoopPC_B.jl
@@ -1,7 +1,0 @@
-module SnoopPC_B
-
-using SnoopPrecompile
-
-@precompile_all_calls missing_function(1)
-
-end # module SnoopPC_B
