--- SnoopPrecompile/test/SnoopPC_C/Project.toml
@@ -1,7 +1,0 @@
-name = "SnoopPC_C"
-uuid = "e6c49816-f2a1-47f5-8743-1735c486fd91"
-authors = ["Tim Holy <tim.holy@gmail.com>"]
-version = "0.1.0"
-
-[deps]
-SnoopPrecompile = "66db9d55-30c0-4569-8b51-7e840670fc0c"
