--- SnoopPrecompile/test/SnoopPC_D/Project.toml
@@ -1,7 +1,0 @@
-name = "SnoopPC_D"
-uuid = "8c347477-c4a5-4b7f-b911-c68b2b9ed891"
-authors = ["Tim Holy <tim.holy@gmail.com>"]
-version = "0.1.0"
-
-[deps]
-SnoopPrecompile = "66db9d55-30c0-4569-8b51-7e840670fc0c"
