+++ test/runtests.jl
@@ -4,3 +4,7 @@ using SnoopCompile
 include("snoopi_deep.jl")
 include("snoopl.jl")
 include("snoopr.jl")
+
+# otherwise-untested demos
+retflat = SnoopCompile.flatten_demo()
+@test !isempty(retflat.children)
